# src/main/resources/swagger.json
{"openapi":"3.0.1","info":{"title":"Inventory Management System","description":"Open API Spec Documentation for Inventory Management System","contact":{"name":"Kayla Dale","url":"https://github.com/KDale119","email":"kdale@mail.mccneb.edu"},"version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/v1/products/{id}":{"get":{"tags":["Product Controller"],"summary":"Returns product by the supplied ID","description":"Get product by ID","operationId":"getProductByID","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"description":"Product JSON Body","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}},"404":{"description":"Not Found"}}},"put":{"tags":["Product Controller"],"summary":"Allows the user to update a product with the supplied data","description":"Update product by ID","operationId":"updateProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"description":"Product JSON Body","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}},"required":true},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}},"404":{"description":"Not Found"},"400":{"description":"Bad Request"}}},"delete":{"tags":["Product Controller"],"summary":"Allows the user to delete a product from inventory","description":"Delete product by ID","operationId":"deleteProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"description":"Product JSON Body","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}},"404":{"description":"Not Found"}}}},"/api/v1/products/{id}/stock":{"put":{"tags":["Product Controller"],"summary":"Allows the user to update only the quantity of a product and leaves the rest of the data the same","description":"Update stock by ID","operationId":"updatedStock","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"description":"Product JSON Body","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}},"required":true},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}},"404":{"description":"Not Found"},"400":{"description":"Bad Request"}}}},"/api/v1/products":{"get":{"tags":["Product Controller"],"summary":"Returns all available products","description":"Get all games","operationId":"getAllProducts","responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}}}}},"post":{"tags":["Product Controller"],"summary":"Allows the user to create a product with the supplied data","description":"Create Product","operationId":"createProduct","requestBody":{"description":"Product JSON Body","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}},"required":true},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}},"400":{"description":"Bad Request"}}}}},"components":{"schemas":{"Product":{"required":["id","name","price"],"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"price":{"maximum":1000,"minimum":0,"type":"number","format":"double"},"quantity":{"maximum":500,"minimum":0,"type":"integer","format":"int32"},"supplier":{"type":"string"},"brand":{"type":"string"}}}}}}